Split the security classes to be in their own files.
